Fix up warnings. Add -Werror to catch new ones.
The lsmsr ones are kind of brute-force fixes because I'm lazy right now.
We should possibly just use unsigned long longs everywhere, instead
of all those u_int64_t's.

+++ lsmsr.c
@@ -165,7 +165,7 @@ do { \
 
 static int _show_msr(struct reg_spec *reg)
 {
-	uint64_t val;
+	unsigned long long val;
 	if (!g.list)
 		if (get_msr_val(reg->address, &val))
 			return 1;
